Update install scripts for WSL/Ubuntu.

+++ tmux/install.sh
@@ -1,12 +1,16 @@
 #!/bin/bash
 
-echo '>>> Install Tmux'
+if command -v pacman &>/dev/null; then
+    echo "--- Install TMUX via Pacman ---------"
+    sudo pacman -S --noconfirm --needed tmux
 
-sudo apt-get -y install tmux
+else
+    echo "--- Install TMUX via apt-get ---------"
+    sudo apt-get -y install tmux
 
-echo '>>> Install Tmux Plugin Manager'
-# NOTE: Might need to install plugins with <Prefix> + I
-rm -rf ~/.tmux/plugins/tpm
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+    echo "--- Install TMUX Plugin Manager ---------"
+    # NOTE: Might need to install plugins with <Prefix> + I
+    rm -rf ~/.tmux/plugins/tpm
+    git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
-
+fi
